refactor(template_active_inference): split integration_audit under the line-count gate
integration_audit.py was 948 lines — a WARN, just 2 under the 950-line fail
threshold, so any further addition would turn the gate red. Split into a
behavior-preserving facade + two cohesive builder modules (re-derived against
main's current version, which differs from the pre-AI-PR-#22 layout):
- integration_audit_builders.py (371) — dependency graph, manuscript provenance,
  claims, gate-index builders + shared helpers
- integration_audit_artifacts.py (438) — diffoscope, license, release-notes,
  figure, scope-boundary, evidence-table, semantic-snapshot builders
- integration_audit.py (254) — facade: writers + validator, re-exports every
  moved name so all existing imports resolve unchanged

Pure move + re-export: public API identical (32 names). Verification:
module_line_count_check exit 0 with no WARN; ruff check + format clean;
import-smoke confirms the public surface + package re-exports; 18 tests pass
(test_roadmap_promotion incl. the integration_audit negative-control battery +
test_output_gates).
